Refactor code as per review
