--- Assets/QA/Tests/Android.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e5a792b571609f14899784f714c9188a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
