+++ Packages/com.unity.inputsystem/InputSystem/Devices/Sensor.cs
@@ -1,4 +1,5 @@
 using System;
+using UnityEngine.InputSystem.Android.LowLevel;
 using UnityEngine.InputSystem.Controls;
 using UnityEngine.InputSystem.Layouts;
 using UnityEngine.InputSystem.LowLevel;
@@ -648,6 +649,47 @@ namespace UnityEngine.InputSystem
         protected override void FinishSetup()
         {
             stepCounter = GetChildControl<IntegerControl>("stepCounter");
+            base.FinishSetup();
+        }
+    }
+
+    /// <summary>
+    /// Hinge angle sensor.
+    /// This sensor is usually available on foldable devices.
+    /// </summary>
+    [InputControlLayout(displayName = "Hinge Angle")]
+    public class HingeAngle : Sensor
+    {
+        /// <summary>
+        /// The angle in degrees on how much the device is unfolded
+        /// </summary>
+        /// <value>0 means fully folded, 180 means fully unfolded.</value>
+        public AxisControl angle { get; protected set; }
+
+        /// <summary>
+        /// The hinge angle sensor that was last added or had activity last.
+        /// </summary>
+        /// <value>Current hinge angle sensor or <c>null</c>.</value>
+        public static HingeAngle current { get; private set; }
+
+        /// <inheritdoc />
+        public override void MakeCurrent()
+        {
+            base.MakeCurrent();
+            current = this;
+        }
+
+        /// <inheritdoc />
+        protected override void OnRemoved()
+        {
+            base.OnRemoved();
+            if (current == this)
+                current = null;
+        }
+
+        protected override void FinishSetup()
+        {
+            angle = GetChildControl<AxisControl>("angle");
             base.FinishSetup();
         }
     }
