+++ Packages/com.unity.inputsystem/InputSystem/Plugins/Android/AndroidSensors.cs
@@ -286,40 +286,8 @@ namespace UnityEngine.InputSystem.Android
     /// </summary>
     /// <seealso href="https://developer.android.com/reference/android/hardware/Sensor#TYPE_HINGE_ANGLE/>
     [InputControlLayout(stateType = typeof(AndroidSensorState), variants = nameof(AndroidSensorType.HingeAngle), hideInUI = true)]
-    public class AndroidHingeAngle : Sensor
+    public class AndroidHingeAngle : HingeAngle
     {
-        /// <summary>
-        /// The angle in degrees on how much the phone is unfolded
-        /// </summary>
-        /// <value>0 means fully folded, 180 means fully unfolded.</value>
-        public AxisControl angle { get; protected set; }
-
-        /// <summary>
-        /// The hinge angle sensor that was last added or had activity last.
-        /// </summary>
-        /// <value>Current hinge angle sensor or <c>null</c>.</value>
-        public static AndroidHingeAngle current { get; private set; }
-
-        /// <inheritdoc />
-        public override void MakeCurrent()
-        {
-            base.MakeCurrent();
-            current = this;
-        }
-
-        /// <inheritdoc />
-        protected override void OnRemoved()
-        {
-            base.OnRemoved();
-            if (current == this)
-                current = null;
-        }
-
-        protected override void FinishSetup()
-        {
-            angle = GetChildControl<AxisControl>("angle");
-            base.FinishSetup();
-        }
     }
 }
